api/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Object storage
minio==7.2.0

# PDF processing and OCR
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# NLP and AI
spacy==3.7.2
sentence-transformers==2.7.0
torch==2.1.1
transformers==4.35.2

# Web scraping
trafilatura==1.6.4
requests==2.31.0
beautifulsoup4==4.12.2

# Document generation
WeasyPrint==60.2
python-docx==1.1.0
Jinja2==3.1.2

# Data processing
pandas==2.1.3
numpy==1.25.2

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
